Ownership, Manage Listings, User Authorization, and Finishing Touches

+++ app/Http/Controllers/ListingController.php
@@ -40,6 +40,8 @@ class ListingController extends Controller
             $formFields['logo'] = $request->file('logo')->store('logos', 'public');
         }
 
+        $formFields['user_id'] = auth()->id();
+
         Listing::create($formFields);
 
         return redirect('/')->with('message', 'Listing created successfully!');
@@ -50,6 +52,10 @@ class ListingController extends Controller
     }
 
     public function update(Request $request, Listing $listing) {
+        if($listing->user_id != auth()->id()) {
+            abort(403, 'Unauthorized action!');
+        }
+        
         $formFields = $request->validate([
             'title' => 'required',
             'company' => 'required',
@@ -70,7 +76,15 @@ class ListingController extends Controller
     }
 
     public function destroy(Listing $listing) {
+        if($listing->user_id != auth()->id()) {
+            abort(403, 'Unauthorized action!');
+        }
+        
         $listing->delete();
         return redirect('/')->with('message', 'Listing deleted successfully');
+    }
+
+    public function manage() {
+        return view('listings.manage', ['listings' => auth()->user()->listings()->get()]);
     }
 }
